Refactor app layout and add providers
This commit refactors the app layout by adding a new component called Providers. The Providers component wraps the existing layout and includes the necessary providers for the app, such as NextUIProvider and QueryClientProvider. This change improves the organization and modularity of the code.

Also, the commit updates the package.json file to include new dependencies: @nanostores/react, @nextui-org/react, @tanstack/react-query, framer-motion, nanostores, and react-hot-toast. It also updates the tailwind.config.ts file to import the nextui theme and adds a plugin for nextui.

The commit also updates the metadata in the layout.tsx file, changing the title to "Adorador" and the description to "Sistema de Lyrics para Iglesias". This change reflects the purpose of the app.

Overall, this commit enhances the app's layout and adds necessary providers and dependencies for improved functionality.